+++ backend-v3/migrations/0036-add-column-merchant-logo.js
@@ -1,0 +1,13 @@
+'use strict'
+
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    return queryInterface.addColumn('merchant', 'logo', {
+      allowNull: true,
+      type: Sequelize.TEXT
+    })
+  },
+  down: (queryInterface, Sequelize) => {
+    return queryInterface.removeColumn('merchant', 'logo')
+  }
+}